+++ uCOS-III_Wrapper/uCOS-III/os_q.c
@@ -667,7 +667,7 @@ OS_OBJ_QTY  OSQPendAbort (OS_Q    *p_q,
                           OS_OPT   opt,
                           OS_ERR  *p_err)
 {
-    rt_uint32_t pend_q_len;
+    OS_OBJ_QTY abort_tasks = 0;
     rt_thread_t thread;
     
     CPU_SR_ALLOC();
@@ -722,15 +722,15 @@ OS_OBJ_QTY  OSQPendAbort (OS_Q    *p_q,
     
     if(opt&OS_OPT_PEND_ABORT_ALL)
     {
-        rt_ipc_pend_abort_all(&(p_q->Msg.parent.suspend_thread));
+        abort_tasks = rt_ipc_pend_abort_all(&(p_q->Msg.parent.suspend_thread));
     }
     else
     {
         rt_ipc_pend_abort_1(&(p_q->Msg.parent.suspend_thread));
+        abort_tasks = 1;
     }
     
     CPU_CRITICAL_ENTER();
-    pend_q_len = rt_list_len(&(p_q->Msg.parent.suspend_thread));/*获取当前还有多少个任务在等待该消息队列*/
 #if OS_CFG_DBG_EN > 0u
     if(!rt_list_isempty(&(p_q->Msg.parent.suspend_thread)))
     {
@@ -751,7 +751,7 @@ OS_OBJ_QTY  OSQPendAbort (OS_Q    *p_q,
     }
     
     *p_err = OS_ERR_NONE;
-    return pend_q_len;
+    return abort_tasks;
 }
 #endif